fix: keep GraphQL helper outside client class
